backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiomysql
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pypdf
langchain
langchain-community
faiss-cpu
openai
sentence-transformers
python-dotenv
pytest
httpx
email-validator
